requirements.txt
asgiref==3.11.1
Django==6.0.2
django-unfold==0.80.2
djangorestframework==3.17.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
